Disable Chrome headless for Windows
This seems to hang, both on Travis and locally. Unable to get any useful logs to understand. In one case, it seems to hang after we call .complete() on a completer - the code awaiting it never runs. This is expecially weird when the code is Dart running in the test runner and shouldn't be affected by the browser mode 🤷‍♂️.